Add messages fetching to App.tsx and render them in ViewMessage
