+++ App/src/App.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import './App.css';
 import { Chats } from './Chats/Chats';
 import { NewMessage } from './Chats/NewMessage/indext';
@@ -6,9 +7,18 @@ import { profilesContext } from './context';
 import { socket } from './socket';
 
 function App() {
+    const [messages, setMessages] = useState<Array<string>>([]);
+
     socket.on('connect', () => {
         console.info('Hey, connected bro');
     });
+
+    function updateMessages(msg: string) {
+        const newMessages = [...messages, msg];
+        setMessages(newMessages);
+    }
+
+    socket.on('chat message', updateMessages);
 
     const profiles = [
         {
@@ -36,7 +46,7 @@ function App() {
             <main>
                 <Chats />
                 <NewMessage />
-                <ViewMessage />
+                <ViewMessage messages={messages} />
                 <h2 className="welcome__title">
                     <img src="/App/src/assets/wepa.png" alt="icon" />
                     <p>WepaChat</p>
